+++ RELEASE-NOTES.md
@@ -7,6 +7,8 @@
 ## Release 1.0.0-RC2
 
 * #360 RowColumnSet: rename "map" to "merge"
+* #364 Empty series isTrue(..) should return "false"
+* #365 BooleanSeries.eq(..) : inefficiency and potential ClassCastException
 
 ## Release 1.0.0-RC1