+++ RELEASE-NOTES.md
@@ -1,3 +1,7 @@
+## Release 2.0.0
+
+* #361 Introduce aggregating semantics in Exp
+
 ## Release 1.1.0
 
 * #357 Primitive FloatSeries